requirements.txt
pandas
python-dotenv
langchain
langchain-groq
openpyxl